Refactoring

+++ src/main/java/com/vidaloca/skibidi/event/repository/EventRepository.java
@@ -1,7 +1,7 @@
 package com.vidaloca.skibidi.event.repository;
 
-import com.vidaloca.skibidi.model.Event;
+import com.vidaloca.skibidi.event.model.Event;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EventRepository extends CrudRepository<Event,Integer> {
+public interface EventRepository extends CrudRepository<Event,Long> {
 }
